# lib/api-client-react/src/generated/api.schemas.ts/submission.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ARENA API
 * OpenAPI spec version: 1.0.0
 */
import type { SubmissionStatus } from './submissionStatus';

export interface Submission {
  id?: number;
  problemId: number;
  contestId?: number | null;
  userId: number;
  username?: string;
  language: string;
  code: string;
  status: SubmissionStatus;
  executionTime?: number | null;
  memoryUsed?: number | null;
  score?: number | null;
  isVirtual: boolean;
  createdAt: string;
}