# lib/api-client-react/src/generated/api.schemas.ts/team.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ARENA API
 * OpenAPI spec version: 1.0.0
 */
import type { TeamMember } from './teamMember';

export interface Team {
  id?: number;
  name: string;
  description?: string | null;
  memberCount: number;
  maxMembers: number;
  rating: number;
  rank: number;
  contestsWon: number;
  totalContests: number;
  members?: TeamMember[];
  createdAt: string;
}