# lib/api-client-react/src/generated/api.schemas.ts/teamMember.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ARENA API
 * OpenAPI spec version: 1.0.0
 */
import type { TeamMemberRole } from './teamMemberRole';

export interface TeamMember {
  userId: number;
  username: string;
  role: TeamMemberRole;
  rating: number;
}