# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: ARENA API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      parameters:
        - in: query
          name: contestId
          schema:
            type: integer
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: problemId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Submission"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /teams:
    get:
      operationId: listTeams
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}/members:
    post:
      operationId: addTeamMember
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                role:
                  type: string
                  enum: [leader, member]
              required:
                - userId
      responses:
        "200":
          description: Added
...
  /users/{id}:
    get:
      operationId: getUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    Submission:
      type: object
      properties:
        id: { type: integer }
        problemId: { type: integer }
        contestId: { type: integer, nullable: true }
        userId: { type: integer }
        username: { type: string }
        language: { type: string }
        code: { type: string }
        status: { type: string, enum: [pending, accepted, wrong_answer, time_limit_exceeded, memory_limit_exceeded, runtime_error, compilation_error] }
        executionTime: { type: integer, nullable: true }
        memoryUsed: { type: integer, nullable: true }
        score: { type: integer, nullable: true }
        isVirtual: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [problemId, userId, language, code, status, isVirtual, createdAt]
    CreateTeamRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        maxMembers: { type: integer, default: 4 }
      required: [name]
    Team:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        memberCount: { type: integer }
        maxMembers: { type: integer }
        rating: { type: integer }
        rank: { type: integer }
        contestsWon: { type: integer }
        totalContests: { type: integer }
        members: { type: array, items: { $ref: "#/components/schemas/TeamMember" } }
        createdAt: { type: string, format: date-time }
      required: [name, memberCount, maxMembers, rating, rank, contestsWon, totalContests, createdAt]
    TeamMember:
      type: object
      properties:
        userId: { type: integer }
        username: { type: string }
        role: { type: string, enum: [leader, member] }
        rating: { type: integer }
      required: [userId, username, role, rating]
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string, nullable: true }
        email: { type: string, nullable: true }
        rating: { type: integer }
        maxRating: { type: integer }
        rank: { type: integer }
        tier: { type: string, enum: [newbie, pupil, specialist, expert, candidate_master, master, grandmaster, legendary] }
        contestsParticipated: { type: integer }
        problemsSolved: { type: integer }
        totalSubmissions: { type: integer }
        acceptanceRate: { type: number }
        country: { type: string, nullable: true }
        teamId: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [username, rating, maxRating, rank, tier, contestsParticipated, problemsSolved, totalSubmissions, acceptanceRate, createdAt]
...